+++ postfix/src/smtpd/smtpd_chat.c
@@ -183,11 +183,11 @@ void    smtpd_chat_reply(SMTPD_STATE *state, char *format,...)
 
 /* print_line - line_wrap callback */
 
-static void print_line(const char *str, ssize_t len, ssize_t indent, char *context)
+static void print_line(const char *str, int len, int indent, char *context)
 {
     VSTREAM *notice = (VSTREAM *) context;
 
-    post_mail_fprintf(notice, " %*s%.*s", (int) indent, "", (int) len, str);
+    post_mail_fprintf(notice, " %*s%.*s", indent, "", len, str);
 }
 
 /* smtpd_chat_notify - notify postmaster */
