+++ postfix/src/util/vstring_vstream.c
@@ -21,17 +21,17 @@
 /*	int	vstring_get_bound(vp, fp, bound)
 /*	VSTRING	*vp;
 /*	VSTREAM	*fp;
-/*	int	bound;
+/*	ssize_t	bound;
 /*
 /*	int	vstring_get_nonl_bound(vp, fp, bound)
 /*	VSTRING	*vp;
 /*	VSTREAM	*fp;
-/*	int	bound;
+/*	ssize_t	bound;
 /*
 /*	int	vstring_get_null_bound(vp, fp, bound)
 /*	VSTRING	*vp;
 /*	VSTREAM	*fp;
-/*	int	bound;
+/*	ssize_t	bound;
 /* DESCRIPTION
 /*	The routines in this module each read one newline or null-terminated
 /*	string from an input stream. In all cases the result is either the
@@ -126,12 +126,12 @@ int     vstring_get_null(VSTRING *vp, VSTREAM *fp)
 
 /* vstring_get_bound - read line from file, keep newline, up to bound */
 
-int     vstring_get_bound(VSTRING *vp, VSTREAM *fp, int bound)
+int     vstring_get_bound(VSTRING *vp, VSTREAM *fp, ssize_t bound)
 {
     int     c;
 
     if (bound <= 0)
-	msg_panic("vstring_get_bound: invalid bound %d", bound);
+	msg_panic("vstring_get_bound: invalid bound %ld", (long) bound);
 
     VSTRING_RESET(vp);
     while (bound-- > 0 && (c = VSTREAM_GETC(fp)) != VSTREAM_EOF) {
@@ -145,12 +145,12 @@ int     vstring_get_bound(VSTRING *vp, VSTREAM *fp, int bound)
 
 /* vstring_get_nonl_bound - read line from file, strip newline, up to bound */
 
-int     vstring_get_nonl_bound(VSTRING *vp, VSTREAM *fp, int bound)
+int     vstring_get_nonl_bound(VSTRING *vp, VSTREAM *fp, ssize_t bound)
 {
     int     c;
 
     if (bound <= 0)
-	msg_panic("vstring_get_nonl_bound: invalid bound %d", bound);
+	msg_panic("vstring_get_nonl_bound: invalid bound %ld", (long) bound);
 
     VSTRING_RESET(vp);
     while (bound-- > 0 && (c = VSTREAM_GETC(fp)) != VSTREAM_EOF && c != '\n')
@@ -161,12 +161,12 @@ int     vstring_get_nonl_bound(VSTRING *vp, VSTREAM *fp, int bound)
 
 /* vstring_get_null_bound - read null-terminated string from file */
 
-int     vstring_get_null_bound(VSTRING *vp, VSTREAM *fp, int bound)
+int     vstring_get_null_bound(VSTRING *vp, VSTREAM *fp, ssize_t bound)
 {
     int     c;
 
     if (bound <= 0)
-	msg_panic("vstring_get_nonl_bound: invalid bound %d", bound);
+	msg_panic("vstring_get_nonl_bound: invalid bound %ld", (long) bound);
 
     VSTRING_RESET(vp);
     while (bound-- > 0 && (c = VSTREAM_GETC(fp)) != VSTREAM_EOF && c != 0)
