+++ postfix/src/util/vstring_vstream.h
@@ -22,9 +22,9 @@
 extern int vstring_get(VSTRING *, VSTREAM *);
 extern int vstring_get_nonl(VSTRING *, VSTREAM *);
 extern int vstring_get_null(VSTRING *, VSTREAM *);
-extern int vstring_get_bound(VSTRING *, VSTREAM *, int);
-extern int vstring_get_nonl_bound(VSTRING *, VSTREAM *, int);
-extern int vstring_get_null_bound(VSTRING *, VSTREAM *, int);
+extern int vstring_get_bound(VSTRING *, VSTREAM *, ssize_t);
+extern int vstring_get_nonl_bound(VSTRING *, VSTREAM *, ssize_t);
+extern int vstring_get_null_bound(VSTRING *, VSTREAM *, ssize_t);
 
  /*
   * Backwards compatibility for code that still uses the vstring_fgets()
